Commit 1 : Models and repositories for CRUD + new SQL

+++ src/test/java/by/project/trucking_v2/config/TestConfig.java
@@ -19,7 +19,7 @@ public class TestConfig {
         public TestUserService(TestUserRepository repo) {
             super(repo);
         }
-        public User getUserById(Long id) {
+        public User getUserById(Integer id) {
             return super.getUserById(id);
         }
     }
